refactor: update folder naming convention
