+++ src/components/Filters/index.js
@@ -1,4 +1,4 @@
-export { PriceFilter } from './Price Filter/PriceFilter';
-export { CategoryFilter } from './Category Filter/CategoryFilter';
-export { RatingFilter } from './Rating Filter/RatingFilter';
-export { SortByPrice } from './ Sort By Price/SortByPrice';
+export { PriceFilter } from './PriceFilter/PriceFilter';
+export { CategoryFilter } from './CategoryFilter/CategoryFilter';
+export { RatingFilter } from './RatingFilter/RatingFilter';
+export { SortByPrice } from './SortByPrice/SortByPrice';
