+++ src/components/index.js
@@ -2,8 +2,8 @@ export { Navbar } from './Navbar/Navbar';
 export { Search } from './Search/Search';
 export { Wrapper } from './Wrapper/Wrapper';
 export { Hero } from './Hero/Hero';
-export { CategoryCard } from './Category Card/CategoryCard';
+export { CategoryCard } from './CategoryCard/CategoryCard';
 export { Loader } from './Loader/Loader';
-export { ProductCard } from './Product Card/ProductCard';
-export { SmallScreenFilter } from './Filters/Small Screen Filter/SmallScreenFilter';
-export { LargeScreenFilter } from './Filters/Large Screen Filter/LargeScreenFilter';
+export { ProductCard } from './ProductCard/ProductCard';
+export { SmallScreenFilter } from './Filters/SmallScreenFilter/SmallScreenFilter';
+export { LargeScreenFilter } from './Filters/LargeScreenFilter/LargeScreenFilter';
